+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+* Rename group -> chunk
+* Rename groupBy -> chunkBy
+
 ## 2.0.0 (2019-03-07)
 
 * Use Array.from to convert iterables to arrays